pyproject.toml
[tool.poetry]
name = "touchbase-demo"
version = "0.1.0"
description = ""
authors = ["Terry Crapts <tcrapts@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
pandas = "^1.0.3"
requests = "^2.23.0"
watchdog = "^0.10.2"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
